refactor: split codex auth collision checks
