+++ src/codex-auth-api.ts
@@ -1,16 +1,14 @@
-import { existsSync, readFileSync } from "node:fs";
-import { join } from "node:path";
-import os from "node:os";
 import { loadConfig, saveConfig } from "./config";
 import {
   getCodexAccountCredential,
   getValidCodexToken,
   saveCodexAccountCredential,
   removeCodexAccountCredential,
-  listCodexAccountIds,
   TokenRefreshError,
 } from "./codex-account-store";
-import { extractAccountId, decodeJwtPayload, extractEmail } from "./oauth/chatgpt";
+import { checkAccountIdCollision, readCodexTokens } from "./codex-auth-collision";
+export { checkAccountIdCollision, getMainChatgptAccountId } from "./codex-auth-collision";
+import { extractAccountId, decodeJwtPayload } from "./oauth/chatgpt";
 import type { OcxConfig } from "./types";
 
 function jsonResponse(data: unknown, status = 200): Response {
@@ -86,70 +84,6 @@ function expireCodexAuthFlow(flowId: string | null, error = "Login cancelled"): 
   if (!flowId) return;
   codexAuthLoginState.set(flowId, { status: "error", error, doneAt: Date.now() });
   setTimeout(() => codexAuthLoginState.delete(flowId), 30_000);
-}
-
-// H1: read main Codex tokens including id_token for reliable account ID extraction
-function readCodexTokens(): { access_token: string; account_id: string; id_token?: string } | null {
-  try {
-    const codexHome = process.env["CODEX_HOME"] || join(os.homedir(), ".codex");
-    const authPath = join(codexHome, "auth.json");
-    if (!existsSync(authPath)) return null;
-    const j = JSON.parse(readFileSync(authPath, "utf-8")) as {
-      tokens?: { access_token?: string; account_id?: string; id_token?: string };
-    };
-    if (j?.tokens?.access_token) {
-      return {
-        access_token: j.tokens.access_token,
-        account_id: j.tokens.account_id ?? "",
-        id_token: j.tokens.id_token,
-      };
-    }
-  } catch { /* best effort */ }
-  return null;
-}
-
-export function getMainChatgptAccountId(): string | null {
-  const tokens = readCodexTokens();
-  if (!tokens) return null;
-  return extractAccountId(tokens.id_token, tokens.access_token) ?? (tokens.account_id || null);
-}
-
-function getMainChatgptEmail(): string | null {
-  const tokens = readCodexTokens();
-  if (!tokens) return null;
-  return extractEmail(tokens.id_token, tokens.access_token) ?? null;
-}
-
-function normalizedEmail(email: string | undefined | null): string | null {
-  const trimmed = email?.trim().toLowerCase();
-  return trimmed || null;
-}
-
-function poolEmailForId(id: string): string | null {
-  const account = (loadConfig().codexAccounts ?? []).find(a => a.id === id);
-  return normalizedEmail(account?.email);
-}
-
-// H2: shared collision check for import and OAuth paths.
-// Business/Team members can share chatgpt_account_id, so require email match too.
-export function checkAccountIdCollision(
-  chatgptAccountId: string,
-  email?: string | null,
-): { collision: true; reason: string } | { collision: false } {
-  const candidateEmail = normalizedEmail(email);
-  const mainId = getMainChatgptAccountId();
-  const mainEmail = getMainChatgptEmail();
-  if (mainId && mainId === chatgptAccountId && (!candidateEmail || !mainEmail || mainEmail === candidateEmail)) {
-    return { collision: true, reason: "This account is your main Codex login. Use a different account for the pool." };
-  }
-  for (const poolId of listCodexAccountIds()) {
-    const cred = getCodexAccountCredential(poolId);
-    const poolEmail = poolEmailForId(poolId);
-    if (cred && cred.chatgptAccountId === chatgptAccountId && (!candidateEmail || !poolEmail || poolEmail === candidateEmail)) {
-      return { collision: true, reason: `Account is already in the pool (${poolId}).` };
-    }
-  }
-  return { collision: false };
 }
 
 let mainAccountCache: { email: string | null; plan: string | null; quota: { weeklyPercent: number; fiveHourPercent: number } | null; ts: number } | null = null;
